dashboard cleanup
